fix(enums): preserve order in encode (#1233)

+++ openfisca_core/indexed_enums/_enum_type.py
@@ -60,7 +60,7 @@ class EnumType(t.EnumType):
         rec.array([(0, 'OWNER', Housing.OWNER), (1, 'TENANT', Housing.TENAN...)
 
         >>> Housing.indices
-        array([0, 1], dtype=int16)
+        array([0, 1], dtype=uint8)
 
         >>> Housing.names
         array(['OWNER', 'TENANT'], dtype='<U6')
